move join test to join
