+++ ydb/core/kqp/ut/runtime/ya.make
@@ -8,7 +8,6 @@ SRCS(
     kqp_scan_spilling_ut.cpp
     kqp_scan_logging_ut.cpp
     kqp_re2_ut.cpp
-    digest_udf_ut.cpp
 )
 
 PEERDIR(
